+++ GWToolbox/GWToolbox/GuiUtils.cpp
@@ -103,3 +103,15 @@ int GuiUtils::ConvertToUtf8(const wchar_t *str, char *output, size_t max_size) {
 int GuiUtils::ConvertToUtf8(const std::wstring& str, char *output, size_t max_size) {
 	return ConvertToUtf8(str.c_str(), output, max_size);
 }
+
+size_t GuiUtils::wcstostr(char *dest, const wchar_t *src, size_t n) {
+	size_t i;
+    unsigned char *d = (unsigned char *)dest;
+    for (i = 0; i < n; i++) {
+        if (src[i] & ~0x7f)
+            return 0;
+        d[i] = src[i] & 0x7f;
+        if (src[i] == 0) break;
+    }
+    return i;
+}