+++ GWToolbox/GWToolbox/GuiUtils.h
@@ -28,4 +28,6 @@ namespace GuiUtils {
 	// Returns the number of bytes in output.
 	int ConvertToUtf8(const wchar_t *str, char *output, size_t max_size);
 	int ConvertToUtf8(const std::wstring& str, char *output, size_t max_size);
+
+	size_t wcstostr(char *dest, const wchar_t *src, size_t n);
 };