+++ GWToolbox/GWToolbox/Widgets/Minimap/PingsLinesRenderer.h
@@ -4,7 +4,8 @@
 #include <vector>
 #include <map>
 
-#include <GWCA\Packets\StoC.h>
+#include <GWCA\Managers\StoCMgr.h>
+#include <GWCA\Managers\UIMgr.h>
 
 #include "VBuffer.h"
 #include "Timer.h"
@@ -14,12 +15,6 @@
 class PingsLinesRenderer : public VBuffer {
 	const float drawing_scale = 96.0f;
 
-	struct ShortPos {
-		ShortPos() : x(0), y(0) {}
-		ShortPos(short _x, short _y) : x(_x), y(_y) {}
-		short x;
-		short y;
-	};
 	struct DrawingLine {
 		DrawingLine() : start(TIMER_INIT()) {}
 		clock_t start;
@@ -79,10 +74,10 @@ public:
 	bool OnMouseMove(float x, float y);
 	bool OnMouseUp();
 
-	void P046Callback(GW::Packet::StoC::P046 *pak);
-	void P138Callback(GW::Packet::StoC::P138 *pak);
-	void P153Callback(GW::Packet::StoC::P153 *pak);
-	void P221Callback(GW::Packet::StoC::P221 *pak);
+	void P046Callback(GW::Packet::StoC::AgentPinged *pak);
+	void P138Callback(GW::Packet::StoC::CompassEvent *pak);
+	void P153Callback(GW::Packet::StoC::GenericValueTarget *pak);
+	void P221Callback(GW::Packet::StoC::SkillActivate *pak);
 
 	void DrawSettings();
 	void LoadSettings(CSimpleIni* ini, const char* section);
@@ -120,7 +115,7 @@ private:
 	clock_t lastshown;
 	clock_t lastsent;
 	clock_t lastqueued;
-	std::vector<ShortPos> queue;
+	std::vector<GW::UI::CompassPoint> queue;
 
 	Color color_drawings;
 	Color color_shadowstep_line;