+++ docs/history.md
@@ -7,6 +7,47 @@ Previous releases are provided as dll files. In order to use it, you need to use
 
 [AutoIt Launcher](https://raw.githubusercontent.com/HasKha/GWToolboxpp/master/AutoitLauncher/Inject.au3) - Source, requires AutoIt3. Right click -> Save link as...
 
+## Version 2.14
+* [Fix] Fixed toolbox functionality after May 9 2018 game update.
+* [Fix] Fixed bug with ctrl+click & dyes.
+
+[Download](https://github.com/HasKha/GWToolboxpp/releases/download/2.14_Release/GWToolbox.dll)
+
+## Version 2.13
+* [Fix] Fixed toolbox functionality after May 9 2018 game update.
+* [New] The damage report will now print the names of the npcs.
+* [Minor] You can now use iso alpha 2 country codes for district with the command `/tp`.
+* [Minor] You can now target player with the command `/target "Player Name"`. (with quotes)
+
+[Download](https://github.com/HasKha/GWToolboxpp/releases/download/2.13_Release/GWToolbox.dll)
+
+## Version 2.12.1
+* [Fix] Removed Borderless Windowed mode, use official implementation instead :)
+* [Fix] Fixed crash bug on toolbox start.
+
+[Download](https://github.com/HasKha/GWToolboxpp/releases/download/2.12.1_Release/GWToolbox.dll)
+
+## Version 2.12
+* [Fix] Fixed toolbox functionality after April 30 2018 game update.
+* [New] Added an option in the chat filter to match messages with regular expressions.
+* [New] Improved custom agent rendering functionality. 
+* [Fix] Empty lines in the chat filter will not ignore everything any more.
+
+[Download](https://github.com/HasKha/GWToolboxpp/releases/download/2.12_Release/GWToolbox.dll)
+
+## Version 2.11
+* [New] Added the option to set custom color for specific agents in the minimap. 
+* [Fix] Fixed toolbox functionality after April 23 2018 game update.
+* [Fix] Improved overall robustness.
+* [Fix] Improved ctrl+click (move items) behaviors.
+
+[Download](https://github.com/HasKha/GWToolboxpp/releases/download/2.11_Release/GWToolbox.dll)
+
+## Version 2.10
+* [Fix] Fixed toolbox functionality after April 2 2018 game update.
+
+[Download](https://github.com/HasKha/GWToolboxpp/releases/download/2.10_Release/GWToolbox.dll)
+
 ## Version 2.9
 * [Fix] Fixed toolbox functionality after March 23 2018 game update.
 * [New] Added a new Trade chat window containing Kamadan Trade Chat. Powered by `https://kamadan.decltype.org/`.